inline codecs
