+++ src/main/java/com/amannmalik/mcp/prompts/PromptListChangedNotification.java
@@ -1,19 +1,18 @@
 package com.amannmalik.mcp.prompts;
 
 import com.amannmalik.mcp.core.JsonCodec;
-import com.amannmalik.mcp.util.EmptyJsonObjectCodec;
-import jakarta.json.JsonObject;
+import jakarta.json.*;
 
 public record PromptListChangedNotification() {
     public static final JsonCodec<PromptListChangedNotification> CODEC = new JsonCodec<>() {
         @Override
         public JsonObject toJson(PromptListChangedNotification n) {
-            return EmptyJsonObjectCodec.toJsonObject();
+            return JsonValue.EMPTY_JSON_OBJECT;
         }
 
         @Override
         public PromptListChangedNotification fromJson(JsonObject obj) {
-            EmptyJsonObjectCodec.requireEmpty(obj);
+            if (obj != null && !obj.isEmpty()) throw new IllegalArgumentException("unexpected fields");
             return new PromptListChangedNotification();
         }
     };
