+++ src/main/java/com/amannmalik/mcp/resources/ResourceListChangedNotification.java
@@ -1,19 +1,18 @@
 package com.amannmalik.mcp.resources;
 
 import com.amannmalik.mcp.core.JsonCodec;
-import com.amannmalik.mcp.util.EmptyJsonObjectCodec;
-import jakarta.json.JsonObject;
+import jakarta.json.*;
 
 public record ResourceListChangedNotification() {
     public static final JsonCodec<ResourceListChangedNotification> CODEC = new JsonCodec<>() {
         @Override
         public JsonObject toJson(ResourceListChangedNotification n) {
-            return EmptyJsonObjectCodec.toJsonObject();
+            return JsonValue.EMPTY_JSON_OBJECT;
         }
 
         @Override
         public ResourceListChangedNotification fromJson(JsonObject obj) {
-            EmptyJsonObjectCodec.requireEmpty(obj);
+            if (obj != null && !obj.isEmpty()) throw new IllegalArgumentException("unexpected fields");
             return new ResourceListChangedNotification();
         }
     };
