+++ src/main/java/com/amannmalik/mcp/tools/ToolListChangedNotification.java
@@ -1,19 +1,18 @@
 package com.amannmalik.mcp.tools;
 
 import com.amannmalik.mcp.core.JsonCodec;
-import com.amannmalik.mcp.util.EmptyJsonObjectCodec;
-import jakarta.json.JsonObject;
+import jakarta.json.*;
 
 public record ToolListChangedNotification() {
     public static final JsonCodec<ToolListChangedNotification> CODEC = new JsonCodec<>() {
         @Override
         public JsonObject toJson(ToolListChangedNotification n) {
-            return EmptyJsonObjectCodec.toJsonObject();
+            return JsonValue.EMPTY_JSON_OBJECT;
         }
 
         @Override
         public ToolListChangedNotification fromJson(JsonObject obj) {
-            EmptyJsonObjectCodec.requireEmpty(obj);
+            if (obj != null && !obj.isEmpty()) throw new IllegalArgumentException("unexpected fields");
             return new ToolListChangedNotification();
         }
     };
