+++ src/main/java/com/amannmalik/mcp/transport/MetadataServlet.java
@@ -15,7 +15,7 @@ final class MetadataServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException {
         ResourceMetadata meta = new ResourceMetadata(transport.canonicalResource, transport.authorizationServers);
-        JsonObject body = ResourceMetadataCodec.toJsonObject(meta);
+        JsonObject body = ResourceMetadata.CODEC.toJson(meta);
         resp.setStatus(HttpServletResponse.SC_OK);
         resp.setContentType("application/json");
         resp.setCharacterEncoding("UTF-8");
